requirements.txt
python-telegram-bot==20.7
langchain==0.1.0
langchain-community==0.0.20
openai==1.6.1
pypdf==3.17.4
faiss-cpu==1.7.4
tiktoken==0.5.2
requests==2.31.0
google-generativeai==0.7.2
sentence-transformers==3.0.1